feat(app): introduce server generated result views
With this commit we made it so that the server
directly includes the result view markup in the
first response. Only from there angular will
take control for subsequent requests.

This is basically for two things:

1. Speed: It's a much nicer user experience
because we safe one extra roundtrip

2. SEO: We can link to example queries that
directly contain useful html Google and other
search engines can deal with.

Fixes #42