fix(package.json): remove nested package.json
We used to have a seperate package.json
for the server subfolder which makes sense
in a way. However, it's not directly supported
by heroku to have multiple package.json
as they essentially treat it as one project.

It doesn't feel nice to have client and server
so tightly mixed but it's better to accept it
for now.

Fixes #7